64-bit frame counts, so we don't get confused when frame counts pass 2^31 (takes about 13 hours)

+++ stepmania/src/RageSoundManager.cpp
@@ -63,7 +63,7 @@ void RageSoundManager::StopMixing( RageSoundBase *snd )
 	}
 }
 
-int RageSoundManager::GetPosition( const RageSoundBase *snd ) const
+int64_t RageSoundManager::GetPosition( const RageSoundBase *snd ) const
 {
 	return driver->GetPosition(snd);
 }
